mi rendo conto che lui vorrebbe le excepions specifiche per alcune entità inserite dove di pertinenza ma nli abbiamo anche exception custom generiche, diventa un bordello,tutto al più stavo pensando ceh possiamo fare dei package ulteriori per dividerle per pertinenza ma sarà un refactor verso la fine. che ne pensate ?
