+++ src/main/java/com/develhope/spring/exceptions/customExceptions/NoResultsException.java
@@ -1,4 +1,4 @@
-package com.develhope.spring.exception;
+package com.develhope.spring.exceptions.customExceptions;
 
 public class NoResultsException extends RuntimeException{
     public NoResultsException(String message) {
