+++ src/main/java/com/develhope/spring/exceptions/customExceptions/UserNotFoundException.java
@@ -1,4 +1,4 @@
-package com.develhope.spring.exception;
+package com.develhope.spring.exceptions.customExceptions;
 
 public class UserNotFoundException extends RuntimeException {
     public UserNotFoundException(String message) {
