+++ src/main/java/com/develhope/spring/exceptions/customExceptions/VehicleNotFoundException.java
@@ -1,4 +1,4 @@
-package com.develhope.spring.exception;
+package com.develhope.spring.exceptions.customExceptions;
 
 public class VehicleNotFoundException extends RuntimeException {
     public VehicleNotFoundException(String message) {
