+++ src/main/java/com/develhope/spring/order/service/OrderService.java
@@ -2,7 +2,7 @@ package com.develhope.spring.order.service;
 
 import com.develhope.spring.order.entity.Order;
 import com.develhope.spring.order.repository.OrderRepository;
-import com.develhope.spring.exception.OrderNotFoundException;
+import com.develhope.spring.exceptions.customExceptions.OrderNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
